pyproject.toml
[tool.poetry]
name = "Mitchi"
version = "0.1.0"
description = "Mitchi - Smart Research Discussion Chatbot"
authors = ["Saefulloh Maslul <maslulsaefulloh@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
langchain = "0.3.2"
langsmith = "0.1.131"
gradio = "4.44.1"
huggingface-hub = "0.25.1"
poethepoet = "0.29.0"
arxiv = "2.1.3"
groq = "0.11.0"
python-dotenv = "1.0.1"
langchain-groq = "0.2.0"


[tool.poetry.group.dev.dependencies]
pylint = "3.3.1"
black = "24.8.0"
isort = "5.13.2"
autoflake = "2.3.1"
pytest = "8.3.3"
pre-commit = "4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = "python main.py"
autoflake = "autoflake"
isort = "isort"
black = "black"
pylint = "pylint"
